services/service-registry/requirements.txt
# Core dependencies
etcd3>=0.12.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
grpcio-health-checking>=1.59.0
grpcio-reflection>=1.59.0

# Data validation and serialization
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Logging and observability
structlog>=23.1.0
prometheus-client>=0.18.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-grpc>=0.41b0
opentelemetry-exporter-otlp>=1.20.0

# Redis for pheromone integration
redis>=5.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
protobuf>=4.24.0
